+++ release_notes.md
@@ -18,6 +18,7 @@
 ### Internals
 
 * Upgrading to OpenSSL 1.0.1t.
+* Added `util::RingBuffer<T>` container.
 
 ----------------------------------------------