+++ release_notes.md
@@ -4,6 +4,9 @@
 
 * Fix initialization of read-only Groups which are sharing file mappings with
   other read-only Groups for the same path.
+* Fixed issue with Timestamps before the UNIX epoch not being read correctly in
+  the `TransactLogParser`. Rollbacks and advances with such Timestamps would
+  throw a `BadTransactLog` exception. (#1802)
 
 ### Breaking changes