+++ release_notes.md
@@ -27,7 +27,8 @@
 
 ### Enhancements
 
-* Lorem ipsum.
+* Changed Array::destroy_deep to be iterative instead of recursive
+  so that Arrays support nesting beyond the stack size. 
 
 -----------